refactor(FloatingPanelAction): Split in a dedicated file
